attempting ga reduction; broken branch
