+++ lib/foxdot.js
@@ -5,11 +5,12 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const async = require('async')
 const child_process_1 = require('child_process')
 const events_1 = require("events");
-const evolver_1 = require("./evolve.js");
-const parse = require ("./parser");
+const evolver_1 = require("./evolveAPI/evolve");
+const parse = require ("./evolveAPI/parser.js");
 const _ =  require("lodash")
 const {Range, Point} = require("atom")
-
+const viz = require("./visuals/index")
+console.log(viz)
 let currentEvo = []
 
 class FoxDot extends events_1.EventEmitter {
@@ -180,7 +181,6 @@ class FoxDot extends events_1.EventEmitter {
       let evolutions = obj.evolver.applyEvolutions(code);
       if (!evolutions) return;
       let params = obj.evolver.getParams();
-      console.log(evolutions)
       let fName = 'evolve' + '_' + params.source + '_' + params.dest + '()',
         genCode = '\n\n' + '@nextBar' +
       '\n' + 'def ' + fName +':'+ '\n'
@@ -203,19 +203,10 @@ class FoxDot extends events_1.EventEmitter {
       let i = 1;
       let interval =  setInterval(() => {
           let j = funcRange[0].start.row;
+            obj.sendVis(i, 'status')
           // anchor ranges around j = end row (func name evolve_p1_p2)
           while (cloneEditor.buffer.getLines()[j] != fName){
             j++;
-            // if (j>10) {
-            //       let clearId = currentEvo[0]['intervalObj']
-            //       currentEvo.splice(0,1)
-            //       console.log(j, progressRange)
-            //       cloneEditor.buffer.setTextInRange(progressRange,'#|')
-            //       console.log(currentEvo)
-            //       obj.clearClock()
-            //       clearInterval(clearId)
-            //       break;
-            // }
           }
           funcRange[0].end.row = j+1;
           progressRange = new Range(new Point(j+1, 0), new Point(j+1, 0))
@@ -244,6 +235,10 @@ class FoxDot extends events_1.EventEmitter {
         start: funcRange[0].start.row,
         intervalObj: interval
       })
+    }
+
+    sendVis(i, key) {
+      viz(i, key)
     }
 
     evaluateRanges(ranges, evolve) {
