+++ lib/index.js
@@ -10,9 +10,10 @@ const child_process_1 = require("child_process");
 const autocomplete_1 = __importDefault(require("./autocomplete"));
 const foxdot_1 = require("./foxdot");
 const logging_1 = require("./logging");
-let foxDot;
-let logger;
-let subscriptions;
+let foxDot, logger, subscriptions;
+const { shell } = require('electron')
+const server = require('./visuals/server')
+
 
 function start() {
     autocomplete_1.default.enabled = true;
@@ -130,6 +131,13 @@ function activate() {
           }
           else {
               foxDot.stopevo()
+          }
+        },
+        'evo:vis': (event) => {
+          if(!foxDot)
+            start();
+          else {
+            shell.openExternal('http://localhost:8080')
           }
         }
