--- lib/params.js
@@ -1,54 +1,0 @@
-// JSON database of all instruments and acceptable attribute value getRanges
-// TODO: move to database?
-
-let instruments = ['loop', 'stretch', 'play1', 'play2', 'audioin', 'noise', 'dab', 'varsaw', 'lazer', 'growl', 'bass', 'dirt', 'crunch', 'rave', 'scatter', 'charm', 'bell', 'gong', 'soprano', 'dub', 'viola', 'scratch', 'klank', 'feel', 'glass', 'soft', 'quin', 'pluck', 'spark', 'blip', 'ripple', 'creep', 'orient', 'zap', 'marimba', 'fuzz', 'bug', 'pulse', 'saw', 'snick', 'twang', 'karp', 'arpy', 'nylon', 'donk', 'squish', 'swell', 'razz', 'sitar', 'star', 'jbass', 'sawbass', 'prophet', 'pads', 'pasha', 'ambi', 'space', 'keys', 'dbass', 'sinepad'];
-let attributes =  {
-  'degree':[-16,16,'int'],
-  'oct':[0,8,'int'],
-  'dur':[0,64,'float'],
-  'delay': [0,10,'float'],
-  'blur':[0,60,'float'],
-  'amplify':[0,10,'float'],
-  'scale':[0,0,'string'],
-  'bpm':[0,300,'int'],
-  'sample':[0,16,'int'],
-  'sus':[0,60,'float'],
-  'fmod':[0,100,'float'],
-  'pan':[-1,1,'float'],
-  'amp':[0.05,1,'float'],
-  'midinote':[0,127,'int'],
-  'vib':[0,20,'float'],
-  'vibdepth':[0, 0.1, 'float'],
-  'slide':[-2,2,'float'],
-  'sus':[0,60,'float'],
-  'slidedelay':[0,1,'float'],
-  'slidefrom':[-2,2,'float'],
-  'glide':[-24,24,'int'],
-  'glidedelay':[0,1,'float'],
-  'bend':[-2,2,'float'],
-  'benddelay':[-2,2,'float'],
-  'coarse':[0,64,'int'],
-  'pshift':[-24,24,'int'],
-  'hpf':[0,16000,'float'],
-  'hpr':[0,1,'float'],
-  'lpf':[0,16000,'float'],
-  'lpr':[0,1,'float'],
-  'bpf':[0,16000,'float'],
-  'bpr':[0,1,'float'],
-  'bpnoise':[0,20000, 'float'],
-  'chop':[0,64,'int'],
-  'striate':[0,500,'int'],
-  'tremolo':[0,128,'int'],
-  'echo':[0,200,'float'],
-  'echotime':[0,8,'int'],
-  'spin':[0,32,'int'],
-  'cut':[0,60,'float'],
-  'room':[0,1,'float'],
-  'mix':[0,1,'float'],
-  'formant':[1,7,'float'],
-  'shape':[0,1,'float'],
-  'drive':[0,1,'float']
-}
-// TODO : 'freq', , 'buf', 'env','rate', 'channel','striate','rate', 'swell','beat_dur',
-
-module.exports = {instruments: instruments, attributes: attributes}
