+++ lib/visuals/index.html
@@ -1,0 +1,59 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <!-- <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
+        <script src="/tsneez/bower_components/d3/d3.min.js"></script>
+        <script src="/tsneez/bower_components/randomcolor/randomColor.js"></script>
+
+        <script src="/tsneez/dist/tsneez.js"></script> -->
+        <script src='./index.js'></script>
+        <style>
+            .container {
+                max-width: 800px;
+                margin: 0 auto;
+                padding: 10px;
+                background-color:whitesmoke;
+            }
+            .viewport {
+                width: 800px;
+                height: 600px;
+            }
+            .viewport svg {
+                width: 800px;
+                height: 600px;
+                display: block;
+            }
+            .viewport svg text {
+                pointer-events: none;
+                font-family: 'Lato', Arial;
+                font-weight: 500;
+                font-size: 1em;
+                fill: #444;
+            }
+        </style>
+    </head>
+    <body>
+        <div class="container">
+          <div class="row">
+            <h1> Big Headidadfng </h1>
+            <div class="viewport">
+            </div>
+          </div>
+        </div>
+    </body>
+</html>
+
+<script>
+    // var ws = new WebSocket('ws://localhost:40510');
+    // // event emmited when connected
+    // ws.onopen = function () {
+    //     // sending a send event to websocket server
+    //     ws.send(window)
+    // }
+    // // event emmited when receiving message
+    // ws.onmessage = function (ev) {
+    //     console.log(ev);
+    // }
+    //
+
+</script>
