+++ test2.py
@@ -17,25 +17,9 @@ Clock.bpm = 101
 Samples.addPath('/Users/a2012/Downloads')
 chords = (VI)
 
-p1>>pluck(room=0.1, amp=0.2,lpf=500)
-p2>>pluck(room=0.8, amp=0.5,lpf=5000)
-evolve(p1,p2,lifetime=4,stepSize=0.2, population=5, skipGenerations=30)
-
-@nextBar
-def evolve_p1_p2():
-	p1>>pluck(room=[0.80],amp=[0.40],lpf=[6240.60])
-	p1.amp=0.5
-	p1.degree=chords
-	p1.sus=8
-	p1.dur=8
-evolve_p1_p2()
-###### Evolving from p1 to p2
-
-
-
-
-
-
+p1 >> pads(degree=[4,6,7,2], dur=[0.5,0.5,4,1], amp=0.5, chop=2)
+p2 >> pads(degree=[4,2,1,3], dur=[0.2,0.5,1,4], amp=0.1, chop=0)
+evolve(p1,p2,lifetime=0.5,stepSize=0.1, population=20, skipGenerations=300, mutationAmount=0, crossoverAmount=1)
 
 k1>>klank(room=0.2, slide=0.6, tremolo=0.5, echo=0.4, bend = 0.5, spin = 4)
